pyproject.toml
[project]
name = "nanamin-cbz-optimizer"
version = "0.1.0"
description = "A CBZ file optimizer with GUI interface"
requires-python = ">=3.10"
dependencies = [
    "PyQt6>=6.4.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.2.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long
    "N802",  # function name should be lowercase - ignore for Qt method overrides
    "B008",  # do not perform function calls in argument defaults
    "B006",  # do not use mutable data structures for argument defaults
    "PLR0913", # too many arguments to function call
]

[tool.ruff.lint.per-file-ignores]
"src/main.py" = ["N802"]  # Ignore Qt method naming in main.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto" 